requirements.txt
# Core Flask framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Database
pymongo==4.6.0

# Configuration
python-dotenv==1.0.0

# AI/ML Services
google-generativeai==0.3.2

# Speech Processing
speech-recognition==3.10.0
pydub==0.25.1
gTTS==2.4.0

# Audio Processing (Windows compatible)
pyaudio==0.2.11

# Utilities
requests==2.31.00
Flask-CORS==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
google-generativeai==0.3.2
SpeechRecognition
pydub==0.25.1
gTTS==2.4.0
pyaudio==0.2.11
requests==2.31.0
Werkzeug==3.0.1